Adapt sqlcipher tests to new api.
